+++ src/menus/ViewMenus.cpp
@@ -101,10 +101,13 @@ double GetZoomOfPreset( const AudacityProject &project, int preset )
          result = pixelsPerUnit * 1000.0;
          break;
       case kZoomSamples:
-         result = 44100.0 * 5.76;
+         result = 44100.0;
          break;
       case kZoom4To1:
          result = 44100.0 * 4;
+         break;
+      case kZoom144th:
+         result = 44100.0 * 5.76;
          break;
       case kMaxZoom:
          result = ZoomInfo::GetMaxZoom();