+++ src/prefs/TracksPrefs.cpp
@@ -223,6 +223,7 @@ static const std::initializer_list<EnumValueSymbol> choicesZoom{
    { wxT("FiveHundredthsOfSeconds"), XO("500ths of Seconds") },
    { XO("MilliSeconds") },
    { XO("Samples") },
+   { wxT("144thSecond"), XO("1/144th Second") },
    { wxT("FourPixelsPerSample"), XO("4 Pixels per Sample") },
    { wxT("MaxZoom"), XO("Max Zoom") },
 };
@@ -241,6 +242,7 @@ static auto enumChoicesZoom = {
    WaveChannelViewConstants::kZoomMilliSeconds,
    WaveChannelViewConstants::kZoomSamples,
    WaveChannelViewConstants::kZoom4To1,
+   WaveChannelViewConstants::kZoom144th,
    WaveChannelViewConstants::kMaxZoom,
 };